+++ Library/OcAfterBootCompatLib/ServiceOverrides.c
@@ -16,7 +16,7 @@
 #include "BootCompatInternal.h"
 
 #include <Guid/AppleVariable.h>
-#include <Guid/OcVariables.h>
+#include <Guid/OcVariable.h>
 
 #include <IndustryStandard/AppleHibernate.h>