+++ Library/OcFileLib/OpenFile.c
@@ -58,43 +58,27 @@ SafeFileOpen (
 
 EFI_STATUS
 EFIAPI
-OcOpenFileByDevicePath (
-  IN OUT EFI_DEVICE_PATH_PROTOCOL  **FilePath,
-  OUT    EFI_FILE_PROTOCOL         **File,
-  IN     UINT64                    OpenMode,
-  IN     UINT64                    Attributes
+OcOpenFileByRemainingDevicePath (
+  IN  EFI_HANDLE                      FileSystemHandle,
+  IN  CONST EFI_DEVICE_PATH_PROTOCOL  *RemainingDevicePath,
+  OUT EFI_FILE_PROTOCOL               **File,
+  IN  UINT64                          OpenMode,
+  IN  UINT64                          Attributes
   )
 {
   EFI_STATUS                      Status;
-  EFI_HANDLE                      FileSystemHandle;
   EFI_SIMPLE_FILE_SYSTEM_PROTOCOL *FileSystem;
   EFI_FILE_PROTOCOL               *LastFile;
-  FILEPATH_DEVICE_PATH            *FilePathNode;
+  CONST EFI_DEVICE_PATH_PROTOCOL  *FilePathNode;
   CHAR16                          *AlignedPathName;
   CHAR16                          *PathName;
   UINTN                           PathLength;
   EFI_FILE_PROTOCOL               *NextFile;
 
-  if (File == NULL) {
-    return EFI_INVALID_PARAMETER;
-  }
-  *File = NULL;
+  ASSERT (FileSystemHandle != NULL);
+  ASSERT (RemainingDevicePath != NULL);
+  ASSERT (File != NULL);
 
-  if (FilePath == NULL) {
-    return EFI_INVALID_PARAMETER;
-  }
-
-  //
-  // Look up the filesystem.
-  //
-  Status = gBS->LocateDevicePath (
-                  &gEfiSimpleFileSystemProtocolGuid,
-                  FilePath,
-                  &FileSystemHandle
-                  );
-  if (EFI_ERROR (Status)) {
-    return Status;
-  }
   Status = gBS->OpenProtocol (
                   FileSystemHandle,
                   &gEfiSimpleFileSystemProtocolGuid,
@@ -119,13 +103,13 @@ OcOpenFileByDevicePath (
   //
   // Traverse the device path nodes relative to the filesystem.
   //
-  while (!IsDevicePathEnd (*FilePath)) {
-    if (DevicePathType (*FilePath) != MEDIA_DEVICE_PATH ||
-        DevicePathSubType (*FilePath) != MEDIA_FILEPATH_DP) {
+  FilePathNode = RemainingDevicePath;
+  while (!IsDevicePathEnd (FilePathNode)) {
+    if (DevicePathType (FilePathNode) != MEDIA_DEVICE_PATH ||
+        DevicePathSubType (FilePathNode) != MEDIA_FILEPATH_DP) {
       Status = EFI_INVALID_PARAMETER;
       goto CloseLastFile;
     }
-    FilePathNode = (FILEPATH_DEVICE_PATH *)*FilePath;
 
     //
     // FilePathNode->PathName may be unaligned, and the UEFI specification
@@ -136,7 +120,7 @@ OcOpenFileByDevicePath (
     AlignedPathName = AllocateCopyPool (
                         (DevicePathNodeLength (FilePathNode) -
                          SIZE_OF_FILEPATH_DEVICE_PATH),
-                        FilePathNode->PathName
+                        ((CONST FILEPATH_DEVICE_PATH *) FilePathNode)->PathName
                         );
     if (AlignedPathName == NULL) {
       Status = EFI_OUT_OF_RESOURCES;
@@ -177,7 +161,7 @@ OcOpenFileByDevicePath (
     //
     LastFile->Close (LastFile);
     LastFile = NextFile;
-    *FilePath = NextDevicePathNode (FilePathNode);
+    FilePathNode = NextDevicePathNode (FilePathNode);
   }
 
   *File = LastFile;
@@ -192,4 +176,40 @@ CloseLastFile:
   //
   ASSERT (EFI_ERROR (Status));
   return Status;
+}
+
+EFI_STATUS
+EFIAPI
+OcOpenFileByDevicePath (
+  IN OUT EFI_DEVICE_PATH_PROTOCOL  **FilePath,
+  OUT    EFI_FILE_PROTOCOL         **File,
+  IN     UINT64                    OpenMode,
+  IN     UINT64                    Attributes
+  )
+{
+  EFI_STATUS Status;
+  EFI_HANDLE FileSystemHandle;
+
+  ASSERT (File != NULL);
+  ASSERT (FilePath != NULL);
+
+  //
+  // Look up the filesystem.
+  //
+  Status = gBS->LocateDevicePath (
+                  &gEfiSimpleFileSystemProtocolGuid,
+                  FilePath,
+                  &FileSystemHandle
+                  );
+  if (EFI_ERROR (Status)) {
+    return Status;
+  }
+
+  return OcOpenFileByRemainingDevicePath (
+           FileSystemHandle,
+           *FilePath,
+           File,
+           OpenMode,
+           Attributes
+           );
 }